requirements.txt
fastapi
uvicorn
spacy
pymupdf
python-docx
sentence-transformers
streamlit
python-multipart
requests
